Fix snooze/DST alarm routing (Alarm.index sentinels)
SNOOZED_ALARM_ID / SUMMER_TIME_ALARM_ID were -1/-2, but Alarm.index is a uint8_t
(also used as the wakeup cookie), so they wrapped to 255/254 in storage while
app.c/alarm_window_logic compared them as ints (-1/-2) -> snooze and DST wakeups
never matched and were mis-handled as group 255/254.

Redefine the sentinels to the uint8 values they were already stored as (255/254).
The stored byte is unchanged (no migration) and the int comparisons against the
wakeup cookie now match. Group ids stay below 254.

Verified on emery: app launches, scheduler recognizes the snooze alarm as index
255, home shows the next group alarm. (Full snooze round-trip needs a ~5-min
wait; routing logic now correct.)

+++ src/c/scheduler.c
@@ -31,10 +31,9 @@ static WakeupId schedule(Alarm* alarm, time_t time)
     {
         alarm->wakeup_id = id;
         // Persist the new wakeup id so the app can still query its alarms after
-        // a relaunch. The index tells us where the alarm lives. NB: index is a
-        // uint8_t, so the negative sentinels wrap (SNOOZED_ALARM_ID -> 255,
-        // SUMMER_TIME_ALARM_ID -> 254); everything else is a group id.
-        if(alarm->index == (uint8_t)SNOOZED_ALARM_ID || alarm->index == (uint8_t)SUMMER_TIME_ALARM_ID)
+        // a relaunch. The index tells us where the alarm lives: the snooze/DST
+        // alarms live in AppConfig, every other index is a group id.
+        if(alarm->index == SNOOZED_ALARM_ID || alarm->index == SUMMER_TIME_ALARM_ID)
         {
             config_save();
         } else
